extract checkPasswordTokenController into its own file
