+++ src/server/controllers/changePassword.ts
@@ -19,17 +19,10 @@ import {
   setEncryptedStateCookie,
 } from '@/server/lib/encryptedStateCookie';
 import { ApiError } from '@/server/models/Error';
-import { RoutePaths } from '@/shared/model/Routes';
-
-type ValidPassworUrls = Extract<
-  '/reset-password' | '/set-password' | '/welcome',
-  RoutePaths
->;
-
-type ValidResendEmailUrls = Extract<
-  '/reset' | '/welcome' | '/set-password',
-  RoutePaths
->;
+import {
+  ValidPasswordResendEmailRoutePaths,
+  ValidPasswordRoutePaths,
+} from '@/shared/model/Routes';
 
 const validatePasswordField = (password: string): Array<FieldError> => {
   const errors: Array<FieldError> = [];
@@ -49,76 +42,10 @@ const validatePasswordField = (password: string): Array<FieldError> => {
   return errors;
 };
 
-export const checkResetPasswordTokenController = (
-  setPasswordPagePath: ValidPassworUrls,
-  setPasswordPageTitle: string,
-  resendEmailPagePath: ValidResendEmailUrls,
-  resendEmailPageTitle: string,
-) =>
-  handleAsyncErrors(async (req: Request, res: ResponseWithRequestState) => {
-    let state = res.locals;
-    const { token } = req.params;
-
-    try {
-      const { email, timeUntilTokenExpiry } = await validateToken(
-        token,
-        req.ip,
-      );
-
-      state = deepmerge(state, {
-        pageData: {
-          browserName: getBrowserNameFromUserAgent(req.header('User-Agent')),
-          email,
-          timeUntilTokenExpiry,
-        },
-      });
-
-      // set the encrypted state here, so we can read the email
-      // on the confirmation page
-      setEncryptedStateCookie(res, { email });
-
-      const html = renderer(
-        `${setPasswordPagePath}${'/:token'}`,
-        {
-          requestState: state,
-          pageTitle: setPasswordPageTitle,
-        },
-        { token },
-      );
-      return res.type('html').send(html);
-    } catch (error) {
-      logger.error(`${req.method} ${req.originalUrl}  Error`, error);
-
-      if (setPasswordPagePath === '/welcome') {
-        const { email, passwordSetOnWelcomePage } =
-          readEncryptedStateCookie(req) ?? {};
-        if (passwordSetOnWelcomePage) {
-          state = deepmerge(state, {
-            pageData: {
-              email,
-            },
-          });
-          return res.type('html').send(
-            renderer(`${resendEmailPagePath}${'/complete'}`, {
-              requestState: state,
-              pageTitle: resendEmailPageTitle,
-            }),
-          );
-        }
-      }
-      return res.type('html').send(
-        renderer(`${resendEmailPagePath}${'/resend'}`, {
-          requestState: state,
-          pageTitle: resendEmailPageTitle,
-        }),
-      );
-    }
-  });
-
 export const setPasswordTokenController = (
-  setPasswordPath: ValidPassworUrls,
+  setPasswordPath: ValidPasswordRoutePaths,
   setPasswordPageTitle: string,
-  resendEmailPagePath: ValidResendEmailUrls,
+  resendEmailPagePath: ValidPasswordResendEmailRoutePaths,
   resendEmailPageTitle: string,
   successCallback: (res: ResponseWithRequestState) => void,
 ) =>
